+++ common/experiment/formats/base.py
@@ -9,7 +9,7 @@ from arekit.common.experiment.neutral.annot.three_scale import ThreeScaleNeutral
 from arekit.common.experiment.neutral.annot.two_scale import TwoScaleNeutralAnnotator
 from arekit.common.experiment.scales.three import ThreeLabelScaler
 from arekit.common.experiment.scales.two import TwoLabelScaler
-from arekit.common.parsed_news.collection import ParsedNewsCollection
+from arekit.common.news.parsed.collection import ParsedNewsCollection
 
 logger = logging.getLogger(__name__)