+++ source/ruattitudes/reader.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from arekit.common.experiment.scales.three import ThreeLabelScaler
-from arekit.common.text_object import TextObject
+from arekit.source.ruattitudes.text_object import TextObject
 from arekit.common.text_opinions.base import RefOpinion
 from arekit.source.ruattitudes.news.base import RuAttitudesNews
 from arekit.source.ruattitudes.sentence import RuAttitudesSentence