+++ tests/text/utils.py
@@ -1,0 +1,20 @@
+from arekit.common.entities.base import Entity
+from arekit.common.frame_variants.base import FrameVariant
+from arekit.processing.text.token import Token
+
+
+def terms_to_str(terms):
+    r = []
+    for t in terms:
+        if isinstance(t, unicode):
+            r.append(t)
+        elif isinstance(t, Token):
+            r.append(t.get_token_value())
+        elif isinstance(t, Entity):
+            r.append(u"[{}]".format(t.Value))
+        elif isinstance(t, FrameVariant):
+            r.append(t.get_value())
+        else:
+            r.append(t)
+
+    return r